bam: backport FireBAM integration to v26.09
Port the BAM integration onto Firedancer v26.09 at
d2e92a3b04006322d74d8292b9dc55b0a49a8c57. Preserve v26.09 gRPC
pacing and caller timestamps while retaining BAM receive suppression,
execution feedback, and the motor maximum-batch MTU fix.

Include the reviewed main performance changes from 7d21e18e6a:
compact transaction storage, indexed BAM batch tracking, and allocation
failure recovery and metrics. Preserve full Firedancer ingress timing;
Frankendancer initializes the shared trailer without unused timestamp
caches or copies. Keep all TOML TLS defaults unchanged.

Retain the four published Agave BAM commits at
fafa43c45f29bb12aa6be0a9a09ed4ee3feeadf3 on the unchanged upstream
Agave base e9c2389bf86c71dd58a91a9e5d30cd2e7baa4fa2. Document
recursive FireBAM builds and initialize pinned submodules in branch CI.

+++ src/discof/replay/fd_sched.c
@@ -2689,7 +2689,7 @@ fd_sched_parse_txn( fd_sched_t * sched, fd_sched_block_t * block, fd_sched_alut_
   sched->metrics->txn_parsed_cnt++;
   sched->txn_pool_free_cnt--;
   fd_txn_p_t * txn_p = sched->txn_pool + txn_idx;
-  txn_p->payload_sz  = pay_sz;
+  txn_p->payload_sz  = (ushort)pay_sz;
 
   txn_p->start_shred_idx = (ushort)shred_split( block, block->fec_buf_boff+block->fec_buf_soff );
   txn_p->start_shred_idx = fd_ushort_if( txn_p->start_shred_idx>0U, (ushort)(txn_p->start_shred_idx-1U), txn_p->start_shred_idx );
